fix: mindmap undo/redo
