+++ packages/blocks/src/surface-block/element-model/decorators/convert.ts
@@ -1,0 +1,41 @@
+import type { ElementModel } from '../base.js';
+import { setObjectMeta } from './common.js';
+
+const convertSymbol = Symbol('convert');
+
+/**
+ * The convert decorator is used to convert the property value before it's
+ * set to the Y map.
+ *
+ * Note:
+ * 1. This decorator function will not execute in model initialization.
+ * @param fn
+ * @returns
+ */
+export function convert<T extends ElementModel>(
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  fn: (propValue: any, instance: T) => unknown
+) {
+  return function convertDecorator(target: unknown, prop: string | symbol) {
+    setObjectMeta(convertSymbol, target, prop as string, fn);
+  };
+}
+
+export function getConvertMeta(
+  target: unknown,
+  prop: string | symbol
+): null | ((propValue: unknown, instance: unknown) => unknown) {
+  // @ts-ignore
+  return target[convertSymbol]?.[prop] ?? null;
+}
+
+export function convertProps(
+  target: unknown,
+  propKey: string | symbol,
+  newProp: unknown,
+  receiver: unknown
+) {
+  const convertFn = getConvertMeta(target, propKey as string)!;
+
+  return convertFn ? convertFn(newProp, receiver) : newProp;
+}
