+++ packages/blocks/src/surface-block/middlewares/connector.ts
@@ -1,9 +1,11 @@
 import type { EdgelessModel } from '../../root-block/edgeless/type.js';
 import type { ConnectorElementModel } from '../index.js';
 import { ConnectorPathGenerator } from '../managers/connector-manager.js';
-import type { SurfaceBlockModel } from '../surface-model.js';
+import type { SurfaceBlockModel, SurfaceMiddleware } from '../surface-model.js';
 
-export function connectorMiddleware(surface: SurfaceBlockModel) {
+export const connectorMiddleware: SurfaceMiddleware = (
+  surface: SurfaceBlockModel
+) => {
   const getElementById = (id: string) =>
     surface.getElementById(id) ??
     (surface.doc.getBlockById(id) as EdgelessModel);
@@ -84,4 +86,4 @@ export function connectorMiddleware(surface: SurfaceBlockModel) {
   return () => {
     disposables.forEach(d => d.dispose());
   };
-}
+};
